+++ src/main/java/io/georocket/util/StreamEvent.java
@@ -6,20 +6,20 @@ package io.georocket.util;
  * @author Michel Kraemer
  */
 public class StreamEvent {
-  private final int pos;
+  private final long pos;
 
   /**
    * Constructs a new event
    * @param pos the position in the input stream where the event has occurred
    */
-  public StreamEvent(int pos) {
+  public StreamEvent(long pos) {
     this.pos = pos;
   }
 
   /**
    * @return the position in the input stream where the event has occurred
    */
-  public int getPos() {
+  public long getPos() {
     return pos;
   }
 }
